Refactored the logic for committing deletes.

+++ src/app/status/components/updated-instance-list/updated-instance-list.component.ts
@@ -11,8 +11,8 @@ import { DeletionDialogService } from 'src/app/instance/components/deletion-dial
 import { ACTION_BUTTONS } from 'src/app/core/models/reactome-schema.model';
 import { ActionButton } from 'src/app/schema-view/list-instances/components/list-instances-view/table/instance-list-table/instance-list-table.component';
 import { Observable, from, concatMap, tap, map, EMPTY } from 'rxjs';
-import { CreateDeletedDialogComponent } from 'src/app/instance/components/create-deleted-dialog/create-deleted-dialog.component';
-import { CreateDeletedDialogService } from 'src/app/instance/components/create-deleted-dialog/create-deleted-dialog.service';
+import { CreateDeletedDialogService } from 'src/app/instance/deletion-commit/components/deleted-object-creation-dialog/deleted-object-creation-dialog.service';
+import { CommitDeletedDialogService } from '../../../instance/deletion-commit/components/deleted-object-creation-option-dialog/deleted-object-creation-option-dialog.service';
 
 
 @Component({
@@ -51,9 +51,11 @@ export class UpdatedInstanceListComponent implements OnInit {
     private dataService: DataService,
     private instanceUtilities: InstanceUtilities,
     private deletionDialogService: DeletionDialogService,
-    private createDeletedDialogService: CreateDeletedDialogService) {
-    }
-  
+    private createDeletedDialogService: CreateDeletedDialogService,
+    private commitDeletedDialogService: CommitDeletedDialogService
+  ) {
+  }
+
 
   ngOnInit(): void {
     this.store.select(newInstances()).subscribe((instances) => {
@@ -123,14 +125,30 @@ export class UpdatedInstanceListComponent implements OnInit {
       // After close needs a flag if the user cancels 
       // If there is a Delete instance created, call a dataservice function to submit this inst 
       // reason will be changed to M and replacement instances will be optional 
-      this.createDeletedDialogService.openDialog().afterClosed().subscribe(result => {});
-      this.dataService.delete(instance).subscribe(rtn => {
-        // Have to subscript it. Otherwise, the http call will not be fired
-        this.store.dispatch(DeleteInstanceActions.remove_deleted_instance(instance));
-        this.instanceUtilities.setDeletedDbId(instance.dbId);
-        this.dataService.flagSchemaTreeForReload();
-      });
+      // If a combo of PE or Event and any other type instance, they user loses the choice for inclucing those other
+      //instances in the deleted instance action. 
+
+      if (this.checkIsEventOrPE(instance)) {
+        this.createDeletedInstance([instance]);
+      }
+      else {
+        this.commitDeletedDialogService.openDialog().afterClosed().subscribe(result => {
+          if (result) {
+            this.createDeletedInstance([instance]);
+          }
+          else {
+            // If not PE or event, just commit the instance
+            this.dataService.delete(instance).subscribe(rtn => {
+              // Have to subscript it. Otherwise, the http call will not be fired
+              this.store.dispatch(DeleteInstanceActions.remove_deleted_instance(instance));
+              this.instanceUtilities.setDeletedDbId(instance.dbId);
+              this.dataService.flagSchemaTreeForReload();
+            });
+          }
+        });
+      }
     }
+
     else if (this.updatedInstances.includes(instance)) {
       this.dataService.commit(instance).subscribe(rtn => {
         this.instanceUtilities.processCommit(instance, rtn, this.dataService);
@@ -155,39 +173,81 @@ export class UpdatedInstanceListComponent implements OnInit {
 
   }
 
-  commitDeletedInstances() {
-    this.selectedDeletedInstances.forEach(instance => {
-      this.dataService.delete(instance).subscribe(rtn => {
-        // Have to subscript it. Otherwise, the http call will not be fired
-        this.store.dispatch(DeleteInstanceActions.remove_deleted_instance(instance));
-        this.instanceUtilities.setDeletedDbId(instance.dbId);
-        this.dataService.flagSchemaTreeForReload();
-      });
-    });
-    this.selectedDeletedInstances = [];
-    this.showCheck = false;
-    this.instanceUtilities.clearSelectedInstances(SelectedInstancesList.deletedInstanceList);
+  checkIsEventOrPE(instance: Instance): boolean {
+    if (this.dataService.isEventClass(instance.schemaClassName) || this.dataService.isPhysicalEntityClass(instance.schemaClassName)) {
+      return true;
+    }
+    return false;
   }
+
+  commitDeletedInstances() {
+    // Check the schema class of every instance to see if a Deleted instance needs to be created
+    let needDeleted = this.selectedDeletedInstances.map(instance => { return this.checkIsEventOrPE(instance) });
+    if (needDeleted.includes(true)) {
+      this.createDeletedInstance(this.selectedDeletedInstances);
+    }
+    else {
+      this.commitDeletedDialogService.openDialog().afterClosed().subscribe(result => {
+        if (result) {
+          this.createDeletedInstance(this.selectedDeletedInstances);
+        }
+        else {
+          // Just commit the deleted instances without creating Deleted instance
+          this.selectedDeletedInstances.forEach(instance => {
+            this.dataService.delete(instance).subscribe(rtn => {
+              this.store.dispatch(DeleteInstanceActions.remove_deleted_instance(instance));
+              this.instanceUtilities.setDeletedDbId(instance.dbId);
+              this.dataService.flagSchemaTreeForReload();
+            });
+          });
+          this.selectedDeletedInstances = [];
+          this.showCheck = false;
+          this.instanceUtilities.clearSelectedInstances(SelectedInstancesList.deletedInstanceList);
+        }
+      });
+    }
+    // TODO: else statment 
+  }
+
+  createDeletedInstance(instanceToDelete: Instance[]) {
+    this.createDeletedDialogService.openDialog(this.selectedDeletedInstances).afterClosed().subscribe(result => {
+      // Submit the Deleted instance, the selected instances are handled in this component
+      this.dataService.deleteByDeleted(result!).subscribe(rtn => {
+
+        instanceToDelete.forEach(instance => {
+          this.store.dispatch(DeleteInstanceActions.remove_deleted_instance(instance));
+          this.instanceUtilities.setDeletedDbId(instance.dbId);
+        });
+
+        this.dataService.flagSchemaTreeForReload();
+        // TODO: debug the marked selected instances after deletion 
+        this.selectedDeletedInstances = [];
+        this.showCheck = false;
+        this.instanceUtilities.clearSelectedInstances(SelectedInstancesList.deletedInstanceList);
+      })
+    })
+  }
+
 
   /**
 * Save a single object via the REST API.
 * Backend may persist related objects and return all saved ones.
 */
-  private saveOne(inst: Instance): Observable < Instance[] > {
-      return this.dataService.commit(inst).pipe(
-        tap(rtn => {
-          this.instanceUtilities.processCommit(inst, rtn, this.dataService);
-        }),
-        map(saved => Array.isArray(saved) ? saved : [saved])
-      );
-    }
+  private saveOne(inst: Instance): Observable<Instance[]> {
+    return this.dataService.commit(inst).pipe(
+      tap(rtn => {
+        this.instanceUtilities.processCommit(inst, rtn, this.dataService);
+      }),
+      map(saved => Array.isArray(saved) ? saved : [saved])
+    );
+  }
 
   /**
 * Save a list of instances via REST API (sequentially).
 * The server may persist related objects automatically.
 */
   commitNewInstances() {
-      if(!this.selectedNewInstances || this.selectedNewInstances.length === 0) {
+    if (!this.selectedNewInstances || this.selectedNewInstances.length === 0) {
       return;
     }
     let shells = this.selectedNewInstances.map(inst => this.instanceUtilities.getShellInstance(inst));
